[#6] 🩹 fix : MainPage 코드 SpaceCard 컴포넌트로 이동
