--- src/styles/MainPage.module.css
@@ -1,16 +1,0 @@
-.container {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 20px;
-  }
-  
-  .spaceGrid {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-    gap: 20px;
-  }
-  
-  .heading {
-    text-align: left; 
-    margin-bottom: 20px; 
-  }
