+++ src/styles/SpaceCard.module.css
@@ -1,17 +1,34 @@
-.spaceCard {
-    background-color: #ffffff; /* 박스의 배경색 */
-    border-radius: 12px; /* 둥근 테두리 */
+.container {
+    max-width: 1200px;
+    margin: 0 auto;
+    padding: 20px;
+  }
+  
+  .heading {
+    text-align: left;
+    margin-bottom: 20px;
+  }
+  
+  .spaceGrid {
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
+    gap: 20px;
+  }
+  
+  .spaceCard {
+    background-color: #ffffff;
+    border-radius: 12px;
     overflow: hidden;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1); /* 그림자 효과 */
-    transition: transform 0.3s ease, box-shadow 0.3s ease; /* 마우스 오버 시 효과 추가 */
-}
-
-.spaceCard:hover {
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+    transition: transform 0.3s ease, box-shadow 0.3s ease;
+  }
+  
+  .spaceCard:hover {
     transform: translateY(-5px);
-    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2); /* 마우스 오버 시 그림자 효과 */
-}
-
-.spaceIcon {
+    box-shadow: 0 8px 16px rgba(0, 0, 0, 0.2);
+  }
+  
+  .spaceIcon {
     width: 100%;
     height: 150px;
     display: flex;
@@ -19,21 +36,22 @@
     justify-content: center;
     font-size: 48px;
     color: #3f51b5;
-    background-color: #e8eaf6; /* 아이콘 배경색 */
-    border-bottom: 1px solid #ddd; /* 아이콘과 정보 구분을 위한 하단 테두리 */
-}
-
-.spaceInfo {
+    background-color: #e8eaf6;
+    border-bottom: 1px solid #ddd;
+  }
+  
+  .spaceInfo {
     padding: 15px;
-}
-
-.spaceTitle {
+  }
+  
+  .spaceTitle {
     font-size: 18px;
     font-weight: bold;
     margin-bottom: 5px;
-}
-
-.spaceDescription {
+  }
+  
+  .spaceDescription {
     font-size: 14px;
     color: #666;
-}
+  }
+
